Move api functionality to Weather component
